Initialize plugin directories and files lazily.
Fixed #388.
